requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
audeer==1.20.1
audioread==3.0.0
audmetric==1.2.0
Babel==2.12.1
bce-python-sdk==0.8.83
blinker==1.6.2
boto3==1.26.133
botocore==1.29.133
bottle==0.12.25
Bottleneck==1.3.7
braceexpand==0.1.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
common==0.1.2
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
data==0.4
dataclasses==0.6
datasets==2.12.0
decorator==5.1.1
dill==0.3.4
Distance==0.1.3
dual==0.0.10
dynamo3==0.4.10
editdistance==0.6.2
fairseq==0.10.1
fastapi==0.95.1
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
Flask-Babel==2.0.0
flatbuffers==23.5.9
flywheel==0.5.4
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
funcsigs==1.0.2
future==0.18.3
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.5.4
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==6.0.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
kaldiio==2.18.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.8.1
llvmlite==0.40.0
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mock==5.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.12.2
nara-wpe==0.0.9
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
onnxruntime==1.14.1
OpenCC==1.1.6
opencc-python-reimplemented==0.1.7
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleaudio==1.1.0
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle==2.4.1
paddlepaddle-gpu==2.4.2
paddleslim==2.4.1
paddlespeech==1.4.1
paddlespeech-feat==0.1.0
pandas==1.2.5
parameterized==0.9.0
pathos==0.2.8
pattern-singleton==1.2.0
peewee==3.16.2
Pillow==9.5.0
platformdirs==3.5.1
pooch==1.7.0
portalocker==2.7.0
pox==0.3.2
ppdiffusers==0.14.1
ppft==1.7.6.6
praatio==5.1.1
prettytable==3.7.0
protobuf==3.20.0
prox==0.0.17
pyarrow==12.0.0
pybind11==2.10.4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.15.1
pygtrie==2.5.0
pyparsing==3.0.9
pypinyin==0.44.0
pypinyin-dict==0.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-geoip-python3==1.3
pytz==2023.3
PyWavelets==1.4.1
pyworld==0.3.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
resampy==0.4.2
responses==0.18.0
rich==13.3.5
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
sentencepiece==0.1.99
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soundfile==0.12.1
starlette==0.26.1
swig==4.1.1
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
TextGrid==1.5
threadpoolctl==3.1.0
tifffile==2023.4.12
tight==0.1.0
timer==0.2.2
ToJyutping==0.2.3
tokenizers==0.10.3
torch==1.13.0
torchaudio==0.8.0a0+a751e1d
torchvision==0.9.0
tqdm==4.65.0
transformers==4.16.2
typeguard==2.13.3
typer==0.9.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
urllib3==1.26.15
uvicorn==0.22.0
visualdl==2.4.2
wcwidth==0.2.6
webrtcvad==2.0.10
websocket-client==1.6.1
websockets==11.0.3
wenetruntime==1.13.0
Werkzeug==2.3.4
xxhash==3.2.0
yacs==0.1.8
yarl==1.9.2
zhon==1.1.5
zipp==3.15.0
